refactor: split plugins/proto/{request, response} into separate files
